backend/requirements.txt
django=5.0.6
django-cors-headers=4.4.0
django-rest-framework=0.10
